Add PolicyContext.IsImageAuthenticated
This is intended for building larger chains of trust, e.g. verifying
layers based on data in the manifest/config.  The policy for other
links in the trust chain (e.g. what to trust about the config or how to
verify it) is not defined by this library, and needs to be handled by
the caller.

NOTE: This is not intended as an all-purpose “trust this image” knob; in
particular, PolicyContext.IsRunningImageAllowed should be used for the
“run this image” decision, and the two may return different results!
(e.g. consider prInsecureAcceptAnything.)

The tests for the critical PolicyContext.IsImageAuthenticated and
prSignedBy.isImageAuthenticated mostly duplicate existing
IsRunningImageAllowed tests (but note the different
prInsecureAnythingAllowed treatment!); a simple duplication of a
straightforward test seems better than either not testing one, or
building a non-trivial abstraction.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>