Add image.UnparsedOriginalInstance
This allows using the signature verification code and image parsing on
the results of GetOriginalManifest, as long as the caller is fine with
the layer data being not available.

Note that UnparsedImage.Manifest() enforces the manifest digest from the
reference, if any, on both the original and current manifest, so if the
two differ, one of the two would fail.  In practice this is not really
an issue for us right now, because docker-daemon: is the only transport
where we have a non-trivial GetOriginalManifest, and uploading to
docker-daemon is only possible with a name:tag, so GetOriginalManifest
can return something only for non-digested references.  (Digested
references could be used for images downloaded via (docker pull),
but those wouldn’t have the original manifest avialable.)

But eventually we might need to solve this somehow.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>