WIP: Cache SigningMechanism objects in PolicyContext
This reduces the time spent in signature tests from ~198 to ~37 seconds
(on macOS, where each GPG agent initialization takes at least a second;
Fedora has a patch which notably speeds this up, so it does not make
that much of a difference there).

This will not make a difference for one-shot copy.Image users,
but it can speed up longer-term users who cache a PolicyContext.

FIXME:
- Split into meanningful commits,
- Review test coverage
- Maybe generalize the state caching instead of hard-coding fields?

Signed-off-by: Miloslav Trmač <mitr@redhat.com>